Update translations W24
ca injection on 20141127-174852

es-ES injection on 20141201-175556

ro injection on 20141209-174921

es-ES injection on 20141215-175237

ca injection on 20141217-175039

es-ES injection on 20141219-175124

Update translation W24.
vi injection on 20141125-113818

ca injection on 20141127-174852

es-ES injection on 20141201-175556

ro injection on 20141209-174921

es-ES injection on 20141215-175237

ca injection on 20141217-175039

es-ES injection on 20141219-175124

es-ES injection on 20141222-175225

ca injection on 20141222-175225

en injection on 20150311-174346

pl injection on 20150316-174728

en injection on 20150409-174447

fr injection on 20150409-174447

vi injection on 20150409-174447

ar injection on 20150409-174447

ja injection on 20150409-174447

de injection on 20150409-174447

sv-SE injection on 20150409-174447

it injection on 20150409-174447

pt-BR injection on 20150409-174447

es-ES injection on 20150409-174447

ru injection on 20150409-174447

tr injection on 20150409-174447

zh-CN injection on 20150409-174447

sq injection on 20150409-174447

zh-TW injection on 20150409-174447

cs injection on 20150409-174447

lt injection on 20150409-174447

uk injection on 20150409-174447

no injection on 20150409-174447

el injection on 20150409-174447

pl injection on 20150409-174447

fa injection on 20150409-174447

ro injection on 20150409-174447

ca injection on 20150409-174447

fr injection on 20150413-174830

Activate again Polish.

en injection on 20150430-174502

vi injection on 20150430-174502

en injection on 20150501-174558

fr injection on 20150507-175641

vi injection on 20150507-175641

ar injection on 20150507-175641

ja injection on 20150507-175641

de injection on 20150507-175641

sv-SE injection on 20150507-175641

it injection on 20150507-175641

pt-BR injection on 20150507-175641

es-ES injection on 20150507-175641

ru injection on 20150507-175641

tr injection on 20150507-175641

zh-CN injection on 20150507-175641

sq injection on 20150507-175641

zh-TW injection on 20150507-175641

cs injection on 20150507-175641

lt injection on 20150507-175641

uk injection on 20150507-175641

no injection on 20150507-175641

el injection on 20150507-175641

pl injection on 20150507-175641

fa injection on 20150507-175641

ro injection on 20150507-175641

ca injection on 20150507-175641

en injection on 20150508-175428

fr injection on 20150508-175428

vi injection on 20150511-175655

ar injection on 20150511-175655

ja injection on 20150511-175655

de injection on 20150511-175655

sv-SE injection on 20150511-175655

it injection on 20150511-175655

pt-BR injection on 20150511-175655

es-ES injection on 20150511-175655

ru injection on 20150511-175655

tr injection on 20150511-175655

zh-CN injection on 20150511-175655

sq injection on 20150511-175655

zh-TW injection on 20150511-175655

cs injection on 20150511-175655

lt injection on 20150511-175655

uk injection on 20150511-175655

no injection on 20150511-175655

el injection on 20150511-175655

pl injection on 20150511-175655

fa injection on 20150511-175655

ro injection on 20150511-175655

ca injection on 20150511-175655

en injection on 20150513-175747

fr injection on 20150514-175048

vi injection on 20150514-175048

ar injection on 20150514-175048

ja injection on 20150514-175048

de injection on 20150514-175048

sv-SE injection on 20150514-175048

it injection on 20150514-175048

pt-BR injection on 20150514-175048

es-ES injection on 20150514-175048

ru injection on 20150514-175048

tr injection on 20150514-175048

zh-CN injection on 20150514-175048

sq injection on 20150514-175048

zh-TW injection on 20150514-175048

cs injection on 20150514-175048

lt injection on 20150514-175048

uk injection on 20150514-175048

no injection on 20150514-175048

el injection on 20150514-175048

pl injection on 20150514-175048

fa injection on 20150514-175048

ro injection on 20150514-175048

ca injection on 20150514-175048

fr injection on 20150515-072919

vi injection on 20150515-072919

ar injection on 20150515-072919

ja injection on 20150515-072919

de injection on 20150515-072919

sv-SE injection on 20150515-072919

it injection on 20150515-072919

pt-BR injection on 20150515-072919

es-ES injection on 20150515-072919

ru injection on 20150515-072919

tr injection on 20150515-072919

zh-CN injection on 20150515-072919

sq injection on 20150515-072919

zh-TW injection on 20150515-072919

cs injection on 20150515-072919

lt injection on 20150515-072919

uk injection on 20150515-072919

no injection on 20150515-072919

el injection on 20150515-072919

pl injection on 20150515-072919

fa injection on 20150515-072919

ro injection on 20150515-072919

ca injection on 20150515-072919

fr injection on 20150515-174700

vi injection on 20150515-174700

en injection on 20150525-175512

en injection on 20150526-175230

fr injection on 20150526-175230

vi injection on 20150526-175230

ar injection on 20150526-175230

ja injection on 20150526-175230

de injection on 20150526-175230

sv-SE injection on 20150526-175230

it injection on 20150526-175230

pt-BR injection on 20150526-175230

es-ES injection on 20150526-175230

ru injection on 20150526-175230

tr injection on 20150526-175230

zh-CN injection on 20150526-175230

sq injection on 20150526-175230

zh-TW injection on 20150526-175230

cs injection on 20150526-175230

lt injection on 20150526-175230

uk injection on 20150526-175230

no injection on 20150526-175230

el injection on 20150526-175230

pl injection on 20150526-175230

fa injection on 20150526-175230

ro injection on 20150526-175230

ca injection on 20150526-175230

fr injection on 20150527-175154

vi injection on 20150527-175154

ar injection on 20150527-175154

ja injection on 20150527-175154

de injection on 20150527-175154

sv-SE injection on 20150527-175154

it injection on 20150527-175154

pt-BR injection on 20150527-175154

es-ES injection on 20150527-175154

ru injection on 20150527-175154

tr injection on 20150527-175154

zh-CN injection on 20150527-175154

sq injection on 20150527-175154

zh-TW injection on 20150527-175154

cs injection on 20150527-175154

lt injection on 20150527-175154

uk injection on 20150527-175154

no injection on 20150527-175154

el injection on 20150527-175154

pl injection on 20150527-175154

fa injection on 20150527-175154

ro injection on 20150527-175154

ca injection on 20150527-175154

fr injection on 20150528-174853

pt-BR injection on 20150528-174853

fr injection on 20150610-174949

ar injection on 20150610-174949

ja injection on 20150610-174949

de injection on 20150610-174949

sv-SE injection on 20150610-174949

it injection on 20150610-174949

pt-BR injection on 20150610-174949

ru injection on 20150610-174949

tr injection on 20150610-174949

zh-CN injection on 20150610-174949

sq injection on 20150610-174949

zh-TW injection on 20150610-174949

lt injection on 20150610-174949

no injection on 20150610-174949

pl injection on 20150610-174949

fa injection on 20150610-174949

ca injection on 20150610-174949