SOC-6148 : add CRSF token in web notifications URLs (#365)
In order to block CSRF attack, a CSRF token is added on Web Notifications action URLs. This must be done on front side since the web notifications action URLs are built on front side in some cases (notification received from websocket).
In order to have the CSRF token of the user session available when building the URLs, it has been injected in the page in the global variable eXo.env.portal.csrfToken.