PLF-8447 : "Send me a digest email" on notification settings is not working (#412)
* PLF-8447 : Resume Daily and weekly digest mails jobs (#406)

* PLF-8447 : Resume Daily and weekly digest mails jobs

This commit add an upgrade plugin to resume jobs in environments in which the migration is already done.

* PLF-8447 : Resume Daily and weekly digest mails jobs

Before this fix, the Upgrade Plugin was not launched if the platform is already in version 6.0
In fact, for tribe, this plugin will be not launched.

The fix change the target version for the UP so that it will be started also on 6.0.x

* PLF-8447 : Send Digest mail are not working
Before this fix, when the Upgrade Plugin is applied after a lot af weeks without running,
all pending notification will be treated and sent.
This Fix remove all daily and weekly notifications which are pending.

* Use MailNotificationStorage instead of DAO

* PLF-8447 : manage error. If an error occurs, throw it to relaunch the UP at next startup

* PLF-8447 : limit the digest to version after 5.2.0

* PLF-8447 : the delete digest must be run if oldVersion is 5.2.0

* Update exception management

* Add unit test

* Update file headers